pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testforge"
version = "0.1.0"
description = "Automated math competition test assembly using ML-predicted difficulty and integer programming."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13,<2.0",
  "beautifulsoup4>=4.12,<5.0",
  "duckdb>=1.1,<2.0",
  "fastapi>=0.115,<1.0",
  "pandas>=2.2,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "pulp>=2.9,<3.0",
  "python-dotenv>=1.0,<2.0",
  "requests>=2.32,<3.0",
  "scikit-learn>=1.6,<2.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
